Add GitHub Actions and Pub to the ecosystems list
This adds GitHub Actions and Pub to the documented list of defined ecosystems. The description for Pub is the wording taken directly from their website, but open to feedback if you'd like it to follow a specific pattern.

There are no validations on the ecosystems defined in the schema docs, so the validation JSON has not been updated.