Update $id to use rolling release
Per https://github.com/ossf/osv-schema/pull/132#issuecomment-1489414045, OSV
uses a rolling release that is fully backwards compatible, preventing the use
of SEMVER for versioning.

Updating $id to reflect this, as there will not be a v2 pushed to main, and all
future versions are defined as backwards compatible with older versions.

Signed-off-by: Josh Buker <crypto@joshbuker.com>