Add Haskell ecosystem vulnerability and advisory types (#157)
There is a Haskell Security Response Team which is now in the process of
defining a repository for security advisories, similar to how Rust
handles these. As part of this, we decided to also integrate with OSV
while we set things up. This should close
https://github.com/haskell/security-advisories/issues/4

There are a few more TODOs left in the PR, partially waiting for
https://github.com/haskell/security-advisories/pull/33#pullrequestreview-1473728366,
partially waiting for a first pass over the PR. I'll try to fill in as
much as possible before this can get merged.

---------

Signed-off-by: Mihai Maruseac <mihaimaruseac@google.com>