Tidy up after a pair-programming session with @agd
Revert the interface, on the premise that there's only going to be one
known implementation at this time.

Rename the types, do away with the custom string and map. Move more of
the definitional variable to the same place as the code.

Simply how checks are inventoried by adding another "ALL" collection.

Signed-off-by: Andrew Pollock <apollock@google.com>