Force at least GitHub package names to lowercase
The Go module proxy seems to not support package names with uppercase in
their name. GitHub URLs are known to be case-insensitive, so it's safe
to explicitly lowercase these. I dare say it'll be safe to lowercase
everything, but I wanted to start conservatively for now.

Also treat Go toolchain vulnerabilities the same way as stdlib ones so
they aren't flagged as a non-existent package.