make redhat conversion tool downloadable with Pip (#278)
In order to install the Red Hat converter tool via Pip we need a
setup.py. Adding that here.

This change avoids specifying the runtime dependencies in both Pipenv
and setup.py by requesting users install the module locally with:

`pipenv install -e .`

As specified in the updated README.md

Signed-off-by: Jason Shepherd <jshepher@redhat.com>