feat: create `osvschema` go bindings package (#292)
This migrates some of the "schema" based types that currently live in
`osv-scanner` over to here since they're related to the underlying OSV
schema rather than the scanner specifically.

Notably to help avoid dependency cycles we now have a dedicated
`constants` package whose sole responsibility is to house types and
constants for representing enums defined in the schema with no logic
whatsoever; this will ensure that this package is always a leaf in the
dependency tree, preventing circular dependencies in libraries (which
notably we have now in the current `osv-scanner/models` package).

This also introduces an `ecosystem` package which houses our concept of
an parsed ecosystem as implicitly defined in the spec - that is, a
struct made up of an ecosystem name and an optional suffix. Note the
underlying `Ecosystem` type actually lives in `constants` which might
seem weird at first but 1. prevents cycles as mentioned above, and 2.
avoids the weird `ecosystem.Ecosystem` situation.

Currently this is being introduced as a port from `osv-scanner` - the
fact that there are missing constants will be addressed in a follow-up
pull request, along with other changes such as updating the
`validation/schema.json` and (hopefully) creating some automation to
help keep everything in sync.

---------

Signed-off-by: Gareth Jones <jones258@gmail.com>