feat: add go struct definitions of OSV record (#333)
Moving the struct definitions of the OSV record from
[osv-scanner](https://github.com/google/osv-scanner/blob/b870d62b9987f27ab0212a9fd0eeef8086554699/pkg/models/vulnerability.go)
to this repository, since it makes more sense to live in the `osvschema`
module.

I've also added the new `upstream` field to the struct. It's probably
worth having a discussion on how to make sure these structs remain
in-sync with any future changes to the schema.

---------

Signed-off-by: Michael Kedar <michaelkedar@google.com>
Co-authored-by: Andrew Pollock <andrewpollock@users.noreply.github.com>