Add V8 prefix. (#339)
### Add V8 prefix to OSV-schema.
As per internal discussions, we are adding the "V8" prefix for
policy-based synthetic vulnerability feeds to OSV schema.
The Chromium team has a
[policy](https://chromium.googlesource.com/chromium/src/+/refs/heads/main/docs/security/updates.md)
that states that direct embedders of Chromium packages such as V8 are
responsible for using an up-to-date version of V8 packages. The
definition of that is that the V8 package is at most one week old.

Since these types of packages do not have their own identifiers and
therefore no advisory could be created, enforcement of such policies was
not possible. We now created a synthetic advisory generator that takes
such time-based policies and converts them into regularily updated
advisories that OSV can read:

https://github.com/google/chromium-policy-vulnfeed/blob/main/advisories/V8-advisory.json

Refer to @oliverchang for context.

Signed-off-by: Viviana Sutedjo <viviana@sutedjo.net>