feat: add `vulnerability.proto` and generated bindings (#417)
This PR adds `vulnerability.proto` to closely bind it with the schema,
as well as adds the generated bindings for Go and Python so we don't
need to manually maintain the bindings.

For Go bindings, this replaces the existing `vulnerability.go` and there
will be another PR soon for `constants.go`.

Signed-off-by: Xueqin Cui <cuixq@google.com>