feat: add a script to generate Go constants based on JSON schema (#419)
Previously, the Go constants were maintained manually, which was
error-prone and led to inconsistencies. This PR introduces a new script
that automates the creation of Go constants based on the JSON schema.

The script dynamically discovers all enum definitions in the JSON schema
and generates a corresponding Go type and const block for each.

---------

Signed-off-by: Xueqin Cui <cuixq@google.com>
Signed-off-by: Jess Lowe <jesslowe@google.com>