Add support for the VSIX / Visual Studio Marketplace ecosystem (#357)
We're contributors to the [OSSF Malicious
Package](https://github.com/ossf/malicious-packages) repository; we'd
like to add our findings for VSIX packages but we're missing the
appropriate ecosystem in the schema.

We'd propose adding a `VSIX` ecosystem, or perhaps `Visual Studio
Marketplace` would be more accurate? The name is a bit chunky so `VS
Marketplace` could also work, we're open to ideas and suggestions.

There aren't any VS(C) extension vulnerability advisories as far as
we're aware so that would only be a ecosystem addition for use in
malicious advisories on the OSSF, if that is not objectionable.

A few examples of such malicious extensions

https://www.reversinglabs.com/blog/malicious-helpers-vs-code-extensions-observed-stealing-sensitive-information
https://x.com/ReversingLabs/status/1902355039265411208

---------

Signed-off-by: rhalar <robert.halar@reversinglabs.com>