fix: make json_values of protos snake_case (#442)
We have a bit of a foot-gun with the proto in that we needed to manually
tell the json marshallers to use the proto names (snake_case), instead
of the generated json fields (camelCase).

There aren't many multiple-word fields in the schema, and the json
schema uses snake_case, so I've added `[json_name="x"]` to the proto
file do avoid needing to do this.

I've also noticed we have a `Commit` message type defined in the proto.
I think this is a really old format that's not in osv anymore, so I've
marked it as deprecated.

---------

Signed-off-by: Michael Kedar <michaelkedar@google.com>