feat: add support for Drupal advisory database (#372)
The [Drupal Advisory
Database](https://github.com/DrupalSecurityTeam/drupal-advisory-database)
is a database maintained by Ackama on behalf of the Drupal community
that provides [Drupal security
advisories](https://www.drupal.org/drupal-security-team/security-advisory-process-and-permissions-policy)
in OSV format to allow them to be ingested by osv.dev.

While Drupal packages are installed and managed using Composer, they are
(mostly) sourced from a dedicated repository rather than the Packagist
repository; since these packages are all within the `drupal/` namespace
which is owned by the Drupal community, it's been agreed that it is fine
to still use the Packagist rather than introduce a new one.

This means that existing tools like `osv-scanner` and libraries like
`osv-scalibr` should "just work" with the advisories in this database.

For the database prefix, the community are decided to use `DRUPAL` as
that is straightforward and matches what other has been proposed in
other tools like
[`dependency-track`](https://github.com/DependencyTrack/dependency-track/pull/4515),
which replaces the "SA" prefix used by advisories published on
drupal.org so advisory ids can be easily mapped to their original
advisory by just replacing `DRUPAL` with `SA`.

Discussion on this can be found
[here](https://www.drupal.org/project/drupalorg/issues/3410338#comment-16200707)
and
[here](https://www.drupal.org/project/drupalorg/issues/3410338#comment-16200707)

---------

Signed-off-by: Gareth Jones <3151613+G-Rath@users.noreply.github.com>