feat(linter): Add a `--new-ecosystem` flag to ignore schema/package checks if its a new ecosystem (#456)
The linter gets flooded with 'Does not match pattern' and 'unsupported
ecosystem' errors when running the linter on a new ecosystem or data
source. That's not very helpful, and is a bit of a catch-22.

This should hide those checks when the `--new-ecosystem` flag is
enabled, so that quality issues that are actionable can be addressed
before submission for new data providers.

Signed-off-by: Jess Lowe <jesslowe@google.com>