feat(linter): use the ecosystem suffix as the repository when linting Packagist advisories (#463)
This replaces the need for custom logic to handle Drupal packages while
also showcasing how the new Packagist suffix works, following the same
[logic that Composer
does](https://getcomposer.org/doc/05-repositories.md#composer) to
resolve a repository to its actual packages.

For now I have just added support for the [`metadata-url`
field](https://getcomposer.org/doc/05-repositories.md#metadata-url-available-packages-and-available-package-patterns)
as that is reasonably straightforward and used by the Drupal
repositories.

Sadly, it seems that https://wpackagist.org and
https://asset-packagist.org still use the Composer v1
[`provider-includes` and
`providers-url`](https://getcomposer.org/doc/05-repositories.md#provider-includes-and-providers-url)
fields which seems to be far more complex to support, but as we only
have custom logic to handle the Drupal repository we're not losing
functionality with this change

---------

Signed-off-by: Gareth Jones <3151613+G-Rath@users.noreply.github.com>